Homepage redesign: separate into multiple pages
With the objective to make the homepage more lean, About and Projects
were moved into their own pages.

The navigation bar moved to the right and gained a "Home" link.

The links that used to be part of the header were moved into the main
part of the homepage, and thus are not displayed in other pages.