Fix flickering when about page finishes loading
With auto column widths, the browser briefly changes the width of the
column after loading and computing the widths of the images.

This commit makes the columns have a fixed size based on the Bootstrap
grid, and let the images adjust to fill the available space, constrained
to fit the column width and the same height as before.