Correct homepage margins
There was no space from the end of the contents to the end of the body,
now we give 1rem of breathing space.

On small screens, there was no space between the image and the links,
now 1rem.

Cleaned up unnecessary markup (extra div, m-0 classes on ul) maintaining
appearance.